% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_systems.R
\name{get_all_systems}
\alias{get_all_systems}
\title{Get all systems currently implemented}
\usage{
get_all_systems(language = NULL)
}
\arguments{
\item{language}{(string) The language of the questions to display.
Defaults to "en".}
}
\value{
(list) All currently implemented usability systems
}
\description{
Get all systems currently implemented. If no language is
  supplied here, all systems will be shown (independent of their language).
  Otherwise (if a language is supplied), the result will only show the
  systems implemented for the given language.
}
