% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_toolings.R
\name{get_all_toolings}
\alias{get_all_toolings}
\title{Get all toolings currently implemented}
\usage{
get_all_toolings()
}
\value{
(list) All currently implemented usability toolings
}
\description{
Get all toolings currently implemented
}
