% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sus_score.R
\name{get_sus_mean}
\alias{get_sus_mean}
\title{Calculate the mean SUS score over multiple respondents}
\usage{
get_sus_mean(data)
}
\arguments{
\item{data}{(data.table) The table containing the results for each question
and respondent. This data must be molten - so if your input is still
very wide (with one column per respondent), first apply
`melt_response_data` on the wide data to obtain long data instead.}
}
\value{
(numeric) The mean over all respondents
}
\description{
Takes a table with the results to each question and
  calculates the overall SUS score.
}
\examples{
\dontrun{
sus_data <- data.table::fread(file = "./data/sus_results_demo.csv")
sus_data_melt <- melt_response_data(data = sus_data)
get_sus_mean(data = sus_data_melt)
}
}
