% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sus_questions.R
\name{get_sus_questions}
\alias{get_sus_questions}
\title{Get all SUS questions.}
\usage{
get_sus_questions()
}
\value{
(list) All currently implemented usability toolings
}
\description{
Get all SUS questions
}
