% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sus_score.R
\name{get_sus_score_raw}
\alias{get_sus_score_raw}
\title{Calculate the SUS score for each respondent.}
\usage{
get_sus_score_raw(data)
}
\arguments{
\item{data}{(data.table) The table containing the results for each question
and respondent.}
}
\value{
(data.table) The SUS score for each respondent.
}
\description{
Takes a table with the results to each question and
  calculates the SUS score for each respondent.
}
