% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sus_score.R
\name{get_sus_score_total}
\alias{get_sus_score_total}
\title{Calculate the SUS score}
\usage{
get_sus_score_total(data)
}
\arguments{
\item{data}{(data.table) The table containing the results for each question
and respondent.}
}
\value{
(data.table) The statistics about the results (min, max, mean,
  median, N).
}
\description{
Takes a table with the results to each question and
  calculates the overall SUS score.
}
