% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sus_score.R
\name{get_sus_stats}
\alias{get_sus_stats}
\title{Calculate the SUS score}
\usage{
get_sus_stats(data)
}
\arguments{
\item{data}{(data.table) The table containing the results for each question
and respondent. This data must be molten - so if your input is still
very wide (with one column per respondent), first apply
`melt_response_data` on the wide data to obtain long data instead.}
}
\value{
(data.table) The statistics about the results (min, max, mean,
  median, N).
}
\description{
Takes a table with the results to each question and
  calculates the stats over all respondents.
}
\examples{
\dontrun{
sus_data <- data.table::fread(file = "./data/sus_results_demo.csv")
sus_data_melt <- melt_response_data(data = sus_data)
get_sus_stats(data = sus_data_melt)[["mean"]]
}

}
