% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{melt_response_data}
\alias{melt_response_data}
\title{Convert wide respondence data to long data.}
\usage{
melt_response_data(data)
}
\arguments{
\item{data}{(data.table) The wide table containing the results for each
question and respondent. Rows = Questions, Cols = responders.}
}
\value{
(data.table) The long table with the same information.
}
\description{
Takes a wide table and converts it to a long data table.
}
