% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sus_score.R
\name{get_sus_score_raw}
\alias{get_sus_score_raw}
\title{Calculate the SUS score for each respondent.}
\usage{
get_sus_score_raw(data, keep_respondents = FALSE)
}
\arguments{
\item{data}{(data.table) The table containing the results for each question
and respondent. This data must be molten - so if your input is still
very wide (with one column per respondent), first apply
`melt_response_data` on the wide data to obtain long data instead.}

\item{keep_respondents}{(boolean, default = `FALSE`) Do you want to get
a named list with the keys beeing the respondents from the input data
and the values beeing the actual result for this respondent?}
}
\value{
(data.table) The SUS score for each respondent.
  There must be a column named `id_official` which contains the official
  SUS question numbers from 1 to 10.
}
\description{
Takes a table with the results to each question and
  calculates the SUS score for each respondent.
}
\examples{
\dontrun{
sus_data <- data.table::fread(file = "./data/sus_results_demo.csv")
sus_data_melt <- melt_response_data(data = sus_data)
get_sus_score_raw(data = sus_data_melt)
}

}
