% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_questions.R
\docType{data}
\name{questions}
\alias{questions}
\title{Questions of different usability scoring systems}
\format{
An object of class \code{"data.table"}.
}
\usage{
data(questions)
}
\description{
Questions of different usability scoring systems
}
\keyword{datasets}
